refactor(channel): refactor
1. remove the `role` field from ChannelNew
2. Don't send ChannelOpen to networker inside ChannelNew
3. Don't dec the `num_pending_out_conn` inside ChannelNew

The ChannelNew only concentrate	on performing the "handshake"
process, but it also need to check whether the remote public
key in the neighbors table.

For (2, 3), the caller need to process the error when failed
to establish a new Channel.

More detail, initiator need to dec the `num_pending_out_conn`
no matter whether the trying conn successful or not.
Both initiator and responder need to send a notification to networker
when the first Channel opened.